+++ tests/version-validator.sh
@@ -98,6 +98,9 @@ expect_pass '2.0.0-rc13.5'
 write_fixture '2.0.0-rc13.6' '2.0.0-rc13.6' '1.0.0'
 expect_pass '2.0.0-rc13.6'
 
+write_fixture '2.0.0-rc13.7' '2.0.0-rc13.7' '1.0.0'
+expect_pass '2.0.0-rc13.7'
+
 write_fixture '2.0.0-rc1' '2.0.0-rc1' '2.0.0'
 expect_fail '2.0.0-rc1'
 
@@ -148,6 +151,9 @@ expect_fail '2.0.0-rc13.4'
 
 write_fixture '2.0.0-rc13.5' '2.0.0-rc13.6' '1.0.0'
 expect_fail '2.0.0-rc13.5'
+
+write_fixture '2.0.0-rc13.6' '2.0.0-rc13.7' '1.0.0'
+expect_fail '2.0.0-rc13.6'
 
 printf 'Release version validator tests passed.\n'